change Long to Double

+++ ChangeDetectorServices/change-detector-service/src/main/java/edu/miu/changedetectorservice/service/StandardDeviationService.java
@@ -4,6 +4,6 @@ import org.springframework.stereotype.Service;
 
 public interface StandardDeviationService {
 
-    void findDataChange(Long x);
+    void findDataChange(Double x);
 }
